(BACK-61) refactor: authorize 패키지 구분
