+++ src/main/java/com/mungwithme/common/annotation/authorize/AdminAuthorize.java
@@ -1,4 +1,4 @@
-package com.mungwithme.common.annotation;
+package com.mungwithme.common.annotation.authorize;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
